+++ packages/vitest-preset-default/scripts/setup-browser.js
@@ -1,0 +1,8 @@
+// Run all tests with development tools enabled.
+// eslint-disable-next-line @wordpress/wp-global-usage
+globalThis.SCRIPT_DEBUG = true;
+
+globalThis.tinyMCEPreInit = {
+	baseURL: 'about:blank',
+};
+globalThis.userSettings = { uid: 1 };