+++ test/integration/helpers/wait-for-store-resolvers.js
@@ -26,7 +26,6 @@ export async function waitForStoreResolvers( fn ) {
 		// resolve before we consider the editor initialized is to flush micro tasks,
 		// similar to the approach found in `@testing-library/react`.
 		// https://github.com/callstack/react-native-testing-library/blob/a010ffdbca906615279ecc3abee423525e528101/src/flushMicroTasks.js#L15-L23.
-		// eslint-disable-next-line testing-library/no-unnecessary-act
 		await act( async () => {} );
 
 		return result;